+++ packages/granularity/src/components/GrDropdown/index.ts
@@ -5,6 +5,6 @@ export { default } from './GrDropdown.vue'
 export { default as GrDropdown } from './GrDropdown.vue'
 export { grDropdownConfig } from './config'
 export { grDropdownSafelist } from './safelist'
-export type { GrDropdownProps, GrDropdownTrigger } from './GrDropdown.vue'
+export type { GrDropdownEmits, GrDropdownProps, GrDropdownTrigger } from './GrDropdown.vue'
 export type { GrDropdownWidth } from './grDropdownStyles'
 export type GrDropdownInstance = ComponentExposed<typeof GrDropdownComponent>